+++ www/docs/docusaurus.config.js
@@ -23,6 +23,12 @@ module.exports = {
         systemvars: true, // Set to true if you would rather load all system variables as well (useful for CI purposes)
       },
     ],
+    [
+      "docusaurus-plugin-segment",
+      {
+        apiKey: process.env.SEGMENT_API_KEY || "temp"
+      }
+    ]
   ],
   themeConfig: {
     colorMode: {