+++ www/reference/gatsby-config.js
@@ -67,11 +67,19 @@ module.exports = {
       },
     },
     {
+      resolve: `gatsby-plugin-segment-js`,
+      options: {
+        prodKey: process.env.SEGMENT_API_KEY,
+        devKey: process.env.SEGMENT_API_KEY_DEV,
+        trackPage: true,
+      }
+    },
+    {
       resolve: `gatsby-plugin-sitemap`,
       options: {
         output: '/api/sitemap'
       }
-    }
+    },
     // `gatsby-plugin-preact`,
     // {
     //   resolve: `gatsby-source-openapi-aggregate`,