+++ packages/medusa-plugin-sendgrid/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/medusajs/medusa/compare/medusa-plugin-sendgrid@1.3.0...medusa-plugin-sendgrid@1.3.2) (2022-06-19)
+
+**Note:** Version bump only for package medusa-plugin-sendgrid
+
+
+
+
+
 ## [1.3.1](https://github.com/medusajs/medusa/compare/medusa-plugin-sendgrid@1.3.0...medusa-plugin-sendgrid@1.3.1) (2022-05-31)
 
 **Note:** Version bump only for package medusa-plugin-sendgrid